Add webdriver setup utility
