+++ src/test/kotlin/tests/WebdriverTest.kt
@@ -8,47 +8,24 @@ import org.junit.jupiter.api.AfterAll
 
 import org.openqa.selenium.WebDriver
 import org.openqa.selenium.interactions.Actions
-import org.openqa.selenium.chrome.ChromeDriver
-import org.openqa.selenium.remote.RemoteWebDriver
-import org.openqa.selenium.remote.DesiredCapabilities
 import org.openqa.selenium.By
 
 import org.team5499.dashboard.Dashboard
-import org.team5499.dashboard.Utils
-
-import java.net.URL
 
 class WebdriverTest {
     lateinit var driver: WebDriver
     lateinit var actions: Actions
-    lateinit var addr: String
 
     companion object {
         lateinit var driver: WebDriver
         lateinit var actions: Actions
-        lateinit var addr: String
 
         @BeforeAll
         @JvmStatic
         fun init() {
             Dashboard.start(this, "webdriverConfig.json")
             Dashboard.awaitInitialization()
-            if (System.getenv("TRAVIS") == "true") {
-                addr = "http://127.0.0.1:5800"
-
-                val capabilities = DesiredCapabilities.chrome()
-                driver = RemoteWebDriver(URL("http://127.0.0.1:4444/wd/hub"), capabilities)
-            } else {
-                addr = "localhost:5800"
-                if (Utils.isWindows()) {
-                    System.setProperty("webdriver.chrome.driver", "src\\test\\resources\\webdrivers\\windows\\chromedriver.exe")
-                } else if (Utils.isMac()) {
-                    System.setProperty("webdriver.chrome.driver", "src/test/resources/webdrivers/mac/chromedriver")
-                } else {
-                    System.setProperty("webdriver.chrome.driver", "src/test/resources/webdrivers/linux/chromedriver")
-                }
-                driver = ChromeDriver()
-            }
+            driver = WebdriverSetup.getDriver()
             actions = Actions(driver)
         }
 
@@ -65,12 +42,11 @@ class WebdriverTest {
     fun setup() {
         driver = WebdriverTest.driver
         actions = WebdriverTest.actions
-        addr = WebdriverTest.addr
     }
 
     @Test
     fun navLinkTest() {
-        driver.get("$addr")
+        driver.get("localhost:5800")
         println(driver.title)
         val navlinks = driver.findElements(By.className("nav-link"))
         val navnames = mutableListOf<String>()
@@ -89,7 +65,7 @@ class WebdriverTest {
     @Test
     fun stringWidgetTest() {
         Dashboard.setVariable("TEST", "testvalue")
-        driver.get("$addr/page/widgettest")
+        driver.get("localhost:5800/page/widgettest")
         Thread.sleep(1000)
         val widgets = driver.findElements(By.className("card-body"))
         widgets.forEach({
@@ -114,7 +90,7 @@ class WebdriverTest {
 
     @Test
     fun newPageTest() {
-        driver.get("$addr")
+        driver.get("localhost:5800")
         val newPageButton = driver.findElements(By.className("btn")).find({ it.text == "New Page" })
         newPageButton?.click()
         val pageName = driver.findElement(By.name("pagename"))
